fix: fix user delete by id endpoint when user doesn't exist

+++ app/crud/crud_task.py
@@ -74,6 +74,8 @@ class CrudTask:
         self, db: AsyncSession, id: Union[int, str]
     ) -> Optional[models.Task]:
         task = await self.get_by_id(db=db, id=id)
+        if not task:
+            return None
         await db.delete(task)
         await db.commit()
         return task
